Weapon Prediction : Prediction flags

+++ include/progs.h
@@ -1196,7 +1196,7 @@ typedef struct gedict_s {
 	func_t client_think;
 	float client_thinkindex;
 	float client_ping;
-	int client_predflags;
+	float client_predflags;
 // }
 } gedict_t;
